Refactor: supportディレクトリ名をtoolsに変更
