+++ internal/usecase/user/user_usecase_test.go
@@ -10,8 +10,8 @@ import (
 	mock_user "boilerplate-go/internal/domain/user/mock"
 	"boilerplate-go/internal/observability"
 	mock_tx "boilerplate-go/internal/usecase/interface/tx/mock"
-	"boilerplate-go/internal/usecase/support/paging"
 	"boilerplate-go/internal/usecase/testkit"
+	"boilerplate-go/internal/usecase/tools/paging"
 	"boilerplate-go/pkg/ptr"
 	"boilerplate-go/pkg/uuid"
